Cleanup of un-intetional package tree
