+++ v2/sourcedb-to-spanner/src/main/java/com/google/cloud/teleport/v2/source/reader/io/jdbc/dialectadapter/package-info.java
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2024 Google LLC
+ * Copyright (C) 2024 Google Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
@@ -13,11 +13,6 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.google.cloud.teleport.v2.source.reader.io.exception.exception;
 
-/** Schema Discovery Exception after Failed Retry attempts. */
-public class SchemaDiscoveryRetriesExhaustedException extends RuntimeException {
-  public SchemaDiscoveryRetriesExhaustedException(Throwable cause) {
-    super(cause);
-  }
-}
+/** Dialect adapter for reader. */
+package com.google.cloud.teleport.v2.source.reader.io.jdbc.dialectadapter;
